# docs/mb-api-spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Message Board API",
        "version": "1.0.1"
    },
    "components": {
...
        "responses": {
            "429TooManyRequests": {
                "description": "Too many requests."
            }
        }
    },
    "servers": [
        {
            "url": "http://10.21.75.182:3004/v1",
            "description": "Message Board API Dev Server"
        }
    ],
    "security": [],
    "paths": {
...
        "/login": {
            "description": "User login",
            "post": {
                "summary": "Log in a user",
                "operationId": "loginUser",
                "description": "Log in a user to the system.",
                "requestBody": {
                    "description": "Login data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "description": "The user's username or email.",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "maxLength": 64,
                                        "description": "The user's password.",
                                        "example": "password123"
                                    }
                                },
                                "required": ["username", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully logged in.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "JWT Token",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request. The Login data in the body of the Request is either missing or malformed."
                    },
                    "401": {
                        "description": "Username or password don't match."
                    },
                    "429": {
                        "$ref": "#/components/responses/429TooManyRequests"
                    }
                }
            }
        }
    }
}